SPLIT RepositoryTest into RepositoryPerformanceTest to run independently
